Add files via upload
This is a script to go through a database and output all Fisher 2 Tailed Exact Test values for every field that has only 'Yes No' or 'True False' columns. It's not statistically sound, but is meant to give you an idea of what possible connections one might be able to find upon further research.